[KZL-374] Auth Controller
