+++ src/config/routes.json
@@ -54,6 +54,16 @@
       "name": "validateMyCredentials",
       "route": "/credentials/:strategy/_me/_validate",
       "method": "post"
+    },
+    "getStrategies": {
+      "name": "getStrategies",
+      "route": "/strategies",
+      "method": "get"
+    },
+    "credentialsExist": {
+      "name": "credentialsExist",
+      "route": "/credentials/:strategy/_me/_exists",
+      "method": "get"
     }
   },
   "bulk": {
