+++ tests/KuzzleTest.php
@@ -200,58 +200,6 @@ class KuzzleTest extends \PHPUnit_Framework_TestCase
         }
     }
 
-    public function testCheckToken()
-    {
-        $fakeToken = uniqid();
-        $url = self::FAKE_KUZZLE_HOST;
-
-        $kuzzle = $this
-            ->getMockBuilder('\Kuzzle\Kuzzle')
-            ->setMethods(['emitRestRequest'])
-            ->setConstructorArgs([$url])
-            ->getMock();
-
-        $options = [
-            'requestId' => uniqid()
-        ];
-
-        // mock http request
-        $httpRequest = [
-            'route' => '/_checkToken',
-            'request' => [
-                'action' => 'checkToken',
-                'controller' => 'auth',
-                'volatile' => [],
-                'body' => [
-                    'token' => $fakeToken
-                ],
-                'requestId' => $options['requestId']
-            ],
-            'method' => 'POST',
-            'query_parameters' => []
-        ];
-
-        // mock response
-        $checkTokenResponse = ['now' => time()];
-        $httpResponse = [
-            'error' => null,
-            'result' => $checkTokenResponse
-        ];
-
-        $kuzzle
-            ->expects($this->once())
-            ->method('emitRestRequest')
-            ->with($httpRequest)
-            ->willReturn($httpResponse);
-
-        /**
-         * @var \Kuzzle\Kuzzle $kuzzle
-         */
-        $response = $kuzzle->checkToken($fakeToken, $options);
-
-        $this->assertEquals($checkTokenResponse, $response);
-    }
-
     public function testCreateIndex()
     {
         $url = self::FAKE_KUZZLE_HOST;
@@ -350,54 +298,6 @@ class KuzzleTest extends \PHPUnit_Framework_TestCase
         $response = $kuzzle->getAllStatistics($options);
 
         $this->assertEquals($getAllStatisticsResponse['hits'], $response);
-    }
-
-    public function testGetMyRights()
-    {
-        $url = self::FAKE_KUZZLE_HOST;
-
-        $kuzzle = $this
-            ->getMockBuilder('\Kuzzle\Kuzzle')
-            ->setMethods(['emitRestRequest'])
-            ->setConstructorArgs([$url])
-            ->getMock();
-
-        $options = [
-            'requestId' => uniqid()
-        ];
-
-        // mock http request
-        $httpRequest = [
-            'route' => '/users/_me/_rights',
-            'request' => [
-                'action' => 'getMyRights',
-                'controller' => 'auth',
-                'volatile' => [],
-                'requestId' => $options['requestId']
-            ],
-            'method' => 'GET',
-            'query_parameters' => []
-        ];
-
-        // mock response
-        $getMyRightsResponse = ['hits' => []];
-        $httpResponse = [
-            'error' => null,
-            'result' => $getMyRightsResponse
-        ];
-
-        $kuzzle
-            ->expects($this->once())
-            ->method('emitRestRequest')
-            ->with($httpRequest)
-            ->willReturn($httpResponse);
-
-        /**
-         * @var \Kuzzle\Kuzzle $kuzzle
-         */
-        $response = $kuzzle->getMyRights($options);
-
-        $this->assertEquals($getMyRightsResponse['hits'], $response);
     }
 
     public function testGetServerInfo()
@@ -780,130 +680,6 @@ class KuzzleTest extends \PHPUnit_Framework_TestCase
         $this->assertEquals($listIndexesResponse['indexes'], $response);
     }
 
-    public function testLogin()
-    {
-        $url = self::FAKE_KUZZLE_HOST;
-        $strategy = 'local';
-        $expiresIn = '1h';
-        $credentials = [
-            'username' => 'foo',
-            'password' => 'bar'
-        ];
-
-        $kuzzle = $this
-            ->getMockBuilder('\Kuzzle\Kuzzle')
-            ->setMethods(['emitRestRequest'])
-            ->setConstructorArgs([$url])
-            ->getMock();
-
-        $options = [
-            'requestId' => uniqid()
-        ];
-
-        // mock http request
-        $httpRequest = [
-            'route' => '/_login/' . $strategy,
-            'request' => [
-                'action' => 'login',
-                'controller' => 'auth',
-                'volatile' => [],
-                'body' => [
-                    'username' => $credentials['username'],
-                    'password' => $credentials['password'],
-                ],
-                'requestId' => $options['requestId']
-            ],
-            'method' => 'POST',
-            'query_parameters' => [
-              'expiresIn' => $expiresIn
-            ]
-        ];
-
-        // mock response
-        $loginResponse = ['jwt' => uniqid()];
-        $httpResponse = [
-            'error' => null,
-            'result' => $loginResponse
-        ];
-
-        $kuzzle
-            ->expects($this->once())
-            ->method('emitRestRequest')
-            ->with($httpRequest)
-            ->willReturn($httpResponse);
-
-        /**
-         * @var \Kuzzle\Kuzzle $kuzzle
-         */
-        $response = $kuzzle->login($strategy, $credentials, $expiresIn, $options);
-
-        $this->assertEquals($loginResponse, $response);
-        $this->assertAttributeEquals($loginResponse['jwt'], 'jwtToken', $kuzzle);
-    }
-
-    public function testLoginWithoutStrategy() {
-        $url = self::FAKE_KUZZLE_HOST;
-        $kuzzle = new \Kuzzle\Kuzzle($url);
-
-        try {
-            $kuzzle->login('');
-
-            $this->fail("KuzzleTest::testLoginWithoutStrategy => Should raise an exception");
-        }
-        catch (Exception $e) {
-            $this->assertInstanceOf('InvalidArgumentException', $e);
-        }
-    }
-
-    public function testLogout()
-    {
-        $url = self::FAKE_KUZZLE_HOST;
-
-        $kuzzle = $this
-            ->getMockBuilder('\Kuzzle\Kuzzle')
-            ->setMethods(['emitRestRequest'])
-            ->setConstructorArgs([$url])
-            ->getMock();
-
-        $options = [
-            'requestId' => uniqid()
-        ];
-
-        // mock http request
-        $httpRequest = [
-            'route' => '/_logout',
-            'request' => [
-                'action' => 'logout',
-                'controller' => 'auth',
-                'volatile' => [],
-                'requestId' => $options['requestId']
-            ],
-            'method' => 'POST',
-            'query_parameters' => []
-        ];
-
-        // mock response
-        $logoutResponse = [];
-        $httpResponse = [
-            'error' => null,
-            'result' => $logoutResponse
-        ];
-
-        $kuzzle
-            ->expects($this->once())
-            ->method('emitRestRequest')
-            ->with($httpRequest)
-            ->willReturn($httpResponse);
-
-        /**
-         * @var \Kuzzle\Kuzzle $kuzzle
-         */
-        $response = $kuzzle->logout($options);
-
-        $this->assertEquals($logoutResponse, $response);
-        $this->assertAttributeEquals(null, 'jwtToken', $kuzzle);
-    }
-
     public function testNow()
     {
         $url = self::FAKE_KUZZLE_HOST;
@@ -1193,113 +969,6 @@ class KuzzleTest extends \PHPUnit_Framework_TestCase
             $this->assertInstanceOf('InvalidArgumentException', $e);
             $this->assertEquals('Unable to set auto refresh on index: no index specified', $e->getMessage());
         }
-    }
-
-    public function testUpdateSelf()
-    {
-        $url = self::FAKE_KUZZLE_HOST;
-        $content = ['foo' => 'bar'];
-
-        $kuzzle = $this
-            ->getMockBuilder('\Kuzzle\Kuzzle')
-            ->setMethods(['emitRestRequest'])
-            ->setConstructorArgs([$url])
-            ->getMock();
-
-        $options = [
-            'requestId' => uniqid()
-        ];
-
-        // mock http request
-        $httpRequest = [
-            'route' => '/_updateSelf',
-            'request' => [
-                'action' => 'updateSelf',
-                'controller' => 'auth',
-                'volatile' => [],
-                'body' => $content,
-                'requestId' => $options['requestId']
-            ],
-            'method' => 'PUT',
-            'query_parameters' => []
-        ];
-
-        // mock response
-        $updateSelfResponse = [];
-        $httpResponse = [
-            'error' => null,
-            'result' => $updateSelfResponse
-        ];
-
-        $kuzzle
-            ->expects($this->once())
-            ->method('emitRestRequest')
-            ->with($httpRequest)
-            ->willReturn($httpResponse);
-
-        /**
-         * @var \Kuzzle\Kuzzle $kuzzle
-         */
-        $response = $kuzzle->updateSelf($content, $options);
-
-        $this->assertEquals($updateSelfResponse, $response);
-    }
-
-    public function testWhoAmI()
-    {
-        $url = self::FAKE_KUZZLE_HOST;
-
-        $kuzzle = $this
-            ->getMockBuilder('\Kuzzle\Kuzzle')
-            ->setMethods(['emitRestRequest'])
-            ->setConstructorArgs([$url])
-            ->getMock();
-
-        $options = [
-            'requestId' => uniqid()
-        ];
-
-        // mock http request
-        $httpRequest = [
-            'route' => '/users/_me',
-            'request' => [
-                'action' => 'getCurrentUser',
-                'controller' => 'auth',
-                'volatile' => [],
-                'requestId' => $options['requestId']
-            ],
-            'method' => 'GET',
-            'query_parameters' => []
-        ];
-
-        // mock response
-        $whoAmIResponse = [
-            '_id' => 'alovelace',
-            '_source' => [
-                'profileIds' => ['admin'],
-                'foo' => 'bar'
-            ],
-            '_meta' => []
-        ];
-        $httpResponse = [
-            'error' => null,
-            'result' => $whoAmIResponse
-        ];
-
-        $kuzzle
-            ->expects($this->once())
-            ->method('emitRestRequest')
-            ->with($httpRequest)
-            ->willReturn($httpResponse);
-
-        /**
-         * @var \Kuzzle\Kuzzle $kuzzle
-         */
-        $response = $kuzzle->whoAmI($options);
-
-        $this->assertInstanceOf('\Kuzzle\Security\User', $response);
-        $this->assertAttributeEquals($whoAmIResponse['_id'], 'id', $response);
-        $this->assertAttributeEquals($whoAmIResponse['_source'], 'content', $response);
     }
 
     public function testBuildQueryArgs()
@@ -1994,224 +1663,6 @@ class KuzzleTest extends \PHPUnit_Framework_TestCase
         $kuzzle->emitEvent($event, 'foo', 'bar');
         $this->assertEquals($triggerCount, 2);
         $this->assertEquals(['foo', 'bar'], $triggerArgs[1]);
-    }
-
-    public function testCreateMyCredentials()
-    {
-        $url = self::FAKE_KUZZLE_HOST;
-
-        $kuzzle = $this
-            ->getMockBuilder('\Kuzzle\Kuzzle')
-            ->setMethods(['emitRestRequest'])
-            ->setConstructorArgs([$url])
-            ->getMock();
-
-        $options = [
-            'requestId' => uniqid()
-        ];
-
-        // mock http request
-        $httpRequest = [
-            'route' => '/credentials/local/_me/_create',
-            'request' => [
-                'action' => 'createMyCredentials',
-                'controller' => 'auth',
-                'volatile' => [],
-                'requestId' => $options['requestId'],
-                'body' => ['foo' => 'bar']
-            ],
-            'method' => 'POST',
-            'query_parameters' => []
-        ];
-
-        $httpResponse = [
-            "result" => [
-                "username" => "foo",
-                "kuid" => "42"
-            ]
-        ];
-
-        $kuzzle
-            ->expects($this->once())
-            ->method('emitRestRequest')
-            ->with($httpRequest)
-            ->willReturn($httpResponse);
-
-        $kuzzle->createMyCredentials("local", ["foo"=>"bar"], $options);
-    }
-
-    public function testDeleteMyCredentials()
-    {
-        $url = self::FAKE_KUZZLE_HOST;
-
-        $kuzzle = $this
-            ->getMockBuilder('\Kuzzle\Kuzzle')
-            ->setMethods(['emitRestRequest'])
-            ->setConstructorArgs([$url])
-            ->getMock();
-
-        $options = [
-            'requestId' => uniqid()
-        ];
-
-        // mock http request
-        $httpRequest = [
-            'route' => '/credentials/local/_me',
-            'request' => [
-                'action' => 'deleteMyCredentials',
-                'controller' => 'auth',
-                'volatile' => [],
-                'requestId' => $options['requestId']
-            ],
-            'method' => 'DELETE',
-            'query_parameters' => []
-        ];
-
-        $httpResponse = [
-            "result" => [
-                "acknowledged" => true
-            ]
-        ];
-
-        $kuzzle
-            ->expects($this->once())
-            ->method('emitRestRequest')
-            ->with($httpRequest)
-            ->willReturn($httpResponse);
-
-        $kuzzle->deleteMyCredentials("local", $options);
-    }
-
-    public function testGetMyCredentials()
-    {
-        $url = self::FAKE_KUZZLE_HOST;
-
-        $kuzzle = $this
-            ->getMockBuilder('\Kuzzle\Kuzzle')
-            ->setMethods(['emitRestRequest'])
-            ->setConstructorArgs([$url])
-            ->getMock();
-
-        $options = [
-            'requestId' => uniqid()
-        ];
-
-        // mock http request
-        $httpRequest = [
-            'route' => '/credentials/local/_me',
-            'request' => [
-                'action' => 'getMyCredentials',
-                'controller' => 'auth',
-                'volatile' => [],
-                'requestId' => $options['requestId']
-            ],
-            'method' => 'GET',
-            'query_parameters' => []
-        ];
-
-        $httpResponse = [
-            "result" => [
-                "username" => "foo",
-                "kuid" => "42"
-            ]
-        ];
-
-        $kuzzle
-            ->expects($this->once())
-            ->method('emitRestRequest')
-            ->with($httpRequest)
-            ->willReturn($httpResponse);
-
-        $kuzzle->getMyCredentials("local", $options);
-    }
-
-    public function testUpdateMyCredentials()
-    {
-        $url = self::FAKE_KUZZLE_HOST;
-
-        $kuzzle = $this
-            ->getMockBuilder('\Kuzzle\Kuzzle')
-            ->setMethods(['emitRestRequest'])
-            ->setConstructorArgs([$url])
-            ->getMock();
-
-        $options = [
-            'requestId' => uniqid()
-        ];
-
-        // mock http request
-        $httpRequest = [
-            'route' => '/credentials/local/_me/_update',
-            'request' => [
-                'action' => 'updateMyCredentials',
-                'controller' => 'auth',
-                'volatile' => [],
-                'requestId' => $options['requestId'],
-                'body' => [
-                    'foo' => 'bar'
-                ]
-            ],
-            'method' => 'PUT',
-            'query_parameters' => []
-        ];
-
-        $httpResponse = [
-            "result" => [
-                "username" => "foo",
-                "kuid" => "42"
-            ]
-        ];
-
-        $kuzzle
-            ->expects($this->once())
-            ->method('emitRestRequest')
-            ->with($httpRequest)
-            ->willReturn($httpResponse);
-
-        $kuzzle->updateMyCredentials("local", ['foo' => 'bar'], $options);
-    }
-
-    public function testValidateMyCredentials()
-    {
-        $url = self::FAKE_KUZZLE_HOST;
-
-        $kuzzle = $this
-            ->getMockBuilder('\Kuzzle\Kuzzle')
-            ->setMethods(['emitRestRequest'])
-            ->setConstructorArgs([$url])
-            ->getMock();
-
-        $options = [
-            'requestId' => uniqid()
-        ];
-
-        // mock http request
-        $httpRequest = [
-            'route' => '/credentials/local/_me/_validate',
-            'request' => [
-                'action' => 'validateMyCredentials',
-                'controller' => 'auth',
-                'volatile' => [],
-                'requestId' => $options['requestId'],
-                'body' => [
-                    'foo' => 'bar'
-                ]
-            ],
-            'method' => 'POST',
-            'query_parameters' => []
-        ];
-
-        $httpResponse = [
-            "result" => true
-        ];
-
-        $kuzzle
-            ->expects($this->once())
-            ->method('emitRestRequest')
-            ->with($httpRequest)
-            ->willReturn($httpResponse);
-
-        $kuzzle->validateMyCredentials("local", ['foo' => 'bar'], $options);
     }
 
     public function testAddHeaders()
